botdir/requirements.txt
black==24.4.2
peewee==3.15.4
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.31.0
telebot==0.0.5
virtualenv==20.26.2
